🚚 api: move usecase to src
