+++ api/src/certification/session/domain/usecases/get-supervisor-kit-session-info.js
@@ -1,4 +1,4 @@
-import { UserNotAuthorizedToAccessEntityError } from '../errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../shared/domain/errors.js';
 
 const getSupervisorKitSessionInfo = async function ({
   userId,
